[Ant4C] allow to set PUGIXML_HEADER_ONLY for all CMake script use.
Fixed typos in messages at CMakeLists.txt.

Removed project argument from echo_evaluate_task as not necessary in
this function.

Added exec_get_attributes_and_arguments_for_task function with test.
Changed signature of exec_evaluate_task function.
Detected issues at Posix side for paths with spaces - quoting in this
case generate an error, so removed such command.

Added verbose parameter to the interpreter_evaluate_task function.

Removed project_str and PROJECT_STR_LENGTH from project unit.
Replaced project_is_in_element with xml_get_tag_name that allow to be as
root element not only project tag.
Added verbose to the project_evaluate_task function.

Added test on exec_evaluate_task function via interpreter_evaluate_task
call.

+++ path.c
@@ -738,7 +738,7 @@ uint8_t path_get_directory_for_current_image(struct buffer* path)
 			return 0;
 		}
 
-		if (expected_size < real_size)
+		if (expected_size < (ptrdiff_t)real_size)
 		{
 			continue;
 		}
